test-requirements.txt
pytest~=7.1.0
pytest-cov>=2.8.1
pytest-randomly>=1.2.3
Flask-Testing==0.8.1
openapi_server
